raop: Move metadata to support
Relates to #1255
