+++ pyatv/support/metadata.py
