+++ pyatv/support/variant.py
