+++ tests/fake_device/mrp.py
@@ -9,12 +9,12 @@ from typing import Dict, Optional, Tuple
 from google.protobuf.message import Message as ProtobufMessage
 
 from pyatv import const
-from pyatv.mrp import messages, protobuf, variant
+from pyatv.mrp import messages, protobuf
 from pyatv.mrp.protobuf import CommandInfo_pb2 as cmd
 from pyatv.mrp.protobuf import PlaybackState
 from pyatv.mrp.protobuf import SendCommandResultMessage as scr
 from pyatv.mrp.server_auth import MrpServerAuth
-from pyatv.support import chacha20, log_protobuf
+from pyatv.support import chacha20, log_protobuf, variant
 
 from tests.utils import stub_sleep
